+++ examples/mocha-express/index.js
@@ -1,1 +1,0 @@
-require('./app').start('localhost', 3000);